Split `download` example into modules.
